% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuadraticFunction.R
\name{Quadder}
\alias{Quadratic Solver}
\title{Creates quadratic function that will allow you to solve the quadratic equation with three inputs}
\usage{
quadraticRoots(a, b, c)
}
\description{
Creates quadratic function that will allow you to solve the quadratic equation with three inputs
}
